Assess normality of traits

Assess normality of traits in a data frame.

assess_normality_core(
  raw_data,
  excluded_columns,
  cpus = 1,
  out_prefix = "metapipe",
  plots_dir = getwd(),
  transf_vals = c(2, exp(1), 3, 4, 5, 6, 7, 8, 9, 10),
  alpha = 0.05
)

Arguments

raw_data
Data frame containing the raw data.

excluded_columns
Numeric vector containing the indices of the dataset properties that are non-numeric, excluded columns.

cpus
Number of CPUs to be used in the computation.

out_prefix
Prefix for output files and plots.

plots_dir
Path to the directory where plots should be stored.

transf_vals
Numeric vector with the transformation values.

alpha
Significance level.

Value

Structure containing the normalised data, if a suitable transformation was found, otherwise returns the original data.

Details

The normality of each trait is checked using a Shapiro-Wilk test, under the following hypotheses:

\(H_0:\) the sample comes from a normally distributed population

\(H_1:\) the sample does not come from a normally distributed population

Using a significance level of \(\alpha = 0.05\). If the conclusion is that the sample does not come from a normally distributed population, then a number of transformations are performed, based on the transformation values passed with transf_vals. By default, the following transformation values are used c(2, exp(1), 3, 4, 5, 6, 7, 8, 9, 10) with the logarithmic (log_a(x)), power (x^a), and radical/root (x^(1/a)) functions.

See also

Examples

# \donttest{
# Toy dataset
example_data <- data.frame(ID = c(1,2,3,4,5), 
                           P1 = c("one", "two", "three", "four", "five"), 
                           T1 = rnorm(5), 
                           T2 = rnorm(5))
example_data_normalised <- MetaPipe:::assess_normality_core(example_data, 
                                                            c(1, 2))
knitr::kable(example_data_normalised)

#> 
#> 
#> | index|trait |     values|flag   |transf |transf_val |
#> |-----:|:-----|----------:|:------|:------|:----------|
#> |     1|T1    | -0.8267890|Normal |       |NA         |
#> |     1|T1    | -1.5123997|Normal |       |NA         |
#> |     1|T1    |  0.9353632|Normal |       |NA         |
#> |     1|T1    |  0.1764886|Normal |       |NA         |
#> |     1|T1    |  0.2436855|Normal |       |NA         |
#> |     2|T2    |  1.6235489|Normal |       |NA         |
#> |     2|T2    |  0.1120381|Normal |       |NA         |
#> |     2|T2    | -0.1339970|Normal |       |NA         |
#> |     2|T2    | -1.9100875|Normal |       |NA         |
#> |     2|T2    | -0.2792372|Normal |       |NA         |

# F1 Seedling Ionomics dataset
data(ionomics) # Includes some missing data
ionomics_rev <- MetaPipe::replace_missing(ionomics, 
                                          excluded_columns = c(1, 2),
                                          replace_na =  TRUE)
ionomics_normalised <- 
  MetaPipe:::assess_normality_core(ionomics_rev,
                                   excluded_columns = c(1, 2),
                                   out_prefix = "ionomics",
                                   transf_vals = c(2, exp(1)))
# Show one entry for each of the first ten traits (left to right)
knitr::kable(ionomics_normalised[nrow(ionomics) * c(1:10), ])

#> 
#> 
#> |     | index|trait |       values|flag   |transf |transf_val |
#> |:----|-----:|:-----|------------:|:------|:------|:----------|
#> |217  |     1|Ca44  | 2.249595e+04|Normal |       |NA         |
#> |434  |     2|K39   | 7.097199e+03|Skewed |       |NA         |
#> |651  |     3|P31   | 2.940119e+03|Skewed |       |NA         |
#> |868  |     4|Li7   | 2.053170e-02|Skewed |       |NA         |
#> |1085 |     5|B11   | 4.342285e+00|Normal |log    |2          |
#> |1302 |     6|Na23  | 2.663747e+00|Normal |root   |e          |
#> |1519 |     7|Mg26  | 1.151307e+01|Normal |log    |2          |
#> |1736 |     8|Al27  | 8.406187e+00|Skewed |       |NA         |
#> |1953 |     9|S34   | 1.425524e+03|Skewed |       |NA         |
#> |2170 |    10|Fe54  | 3.870016e+01|Skewed |       |NA         |

# }

# Clean up example outputs
MetaPipe:::tidy_up("HIST")