Post-processing for assess_normality

Post-processing for the normality assessment of the traits.

assess_normality_postprocessing(
  raw_data,
  excluded_columns,
  raw_data_normalised,
  out_prefix = "metapipe",
  pareto_scaling = FALSE
)

Arguments

raw_data
Data frame containing the raw data.

excluded_columns
Numeric vector containing the indices of the dataset properties that are non-numeric, excluded columns.

raw_data_normalised
Data frame containing the normalised raw data, created with assess_normality.

out_prefix
Prefix for output files and plots.

pareto_scaling
Boolean flag to indicate whether or not perform a Pareto scaling on the normalised data.

Value

List of data frames for the normal (norm) and skewed (skew) traits.

Details

It creates four CSV files containing the following data:

paste0(out_prefix, "_raw_data_normalised_all.csv"): data frame in long format created with the assess_normality function. Contains all the traits, including name, values, and transformation (if applicable).

paste0(out_prefix, "_raw_data_norm.csv"): data frame in wide format containing traits with a normal distribution.

paste0(out_prefix, "_raw_data_non_par.csv"): data frame in wide format containing traits without a normal distribution.

paste0(out_prefix, "_normalisation_stats.csv"): data frame containing details of the assess_normality process, like the total number of traits transformed and normalisations performed.

See also

Examples

# Toy dataset
example_data <- data.frame(ID = c(1,2,3,4,5), 
                           P1 = c("one", "two", "three", "four", "five"), 
                           T1 = rnorm(5), 
                           T2 = rnorm(5))
example_data_normalised <- MetaPipe::assess_normality(example_data, c(1, 2))
example_data_normalised_post <- 
 MetaPipe::assess_normality_postprocessing(example_data, 
                                           c(1, 2), 
                                           example_data_normalised)
example_data_norm <- example_data_normalised_post$norm
example_data_skew <- example_data_normalised_post$skew
# Normal traits
knitr::kable(example_data_norm)

#> 
#> 
#> | ID|         T1|         T2|
#> |--:|----------:|----------:|
#> |  1| -0.5536994| -1.8630115|
#> |  2|  0.6289820| -0.5220125|
#> |  3|  2.0650249| -0.0526019|
#> |  4| -1.6309894|  0.5429963|
#> |  5|  0.5124269| -0.9140748|

# Skewed traits (empty)
# knitr::kable(example_data_skew)

# F1 Seedling Ionomics dataset
data(ionomics) # Includes some missing data
ionomics_rev <- MetaPipe::replace_missing(ionomics, 
                                          excluded_columns = c(1, 2),
                                          replace_na =  TRUE)
ionomics_normalised <- 
  MetaPipe::assess_normality(ionomics_rev,
                             excluded_columns = c(1, 2),
                             out_prefix = "ionomics",
                             transf_vals = c(2, exp(1)))
ionomics_normalised_post <- 
  MetaPipe::assess_normality_postprocessing(ionomics_rev, 
                                            c(1, 2), 
                                            ionomics_normalised)
ionomics_norm <- ionomics_normalised_post$norm
ionomics_skew <- ionomics_normalised_post$skew
# Normal traits
knitr::kable(ionomics_norm[1:5, ])

#> 
#> 
#> |ID    |     Ca44|      B11|     Na23|     Mg26|     Rb85|      Sr88|
#> |:-----|--------:|--------:|--------:|--------:|--------:|---------:|
#> |E_001 | 15894.22| 4.222397| 2.042740| 10.77021| 1.555742|  7.347059|
#> |E_002 | 13155.45| 3.855684| 1.917202| 10.54095| 2.058711|  6.890243|
#> |E_004 | 14182.51| 3.879033| 2.354263| 10.51931| 2.198422|  9.025915|
#> |E_005 | 22550.82| 4.329576| 2.477233| 11.13450| 1.791578| 15.292360|
#> |E_006 | 15982.76| 3.697997| 2.419593| 11.72734| 2.229866| 13.901449|

# Skewed traits (partial output)
knitr::kable(ionomics_skew[1:5, 1:10])

#> 
#> 
#> |ID    |      K39|      P31|       Li7|      Al27|      S34|     Fe54|     Mn55|      Co59|      Ni60|
#> |:-----|--------:|--------:|---------:|---------:|--------:|--------:|--------:|---------:|---------:|
#> |E_001 | 5888.311| 1743.118| 0.0128699|  3.845879| 1152.944| 27.59340| 54.53991| 0.1887654| 0.8138975|
#> |E_002 | 7013.400| 2244.684| 0.0119316|  5.825639| 1600.442| 35.49159| 52.57114| 0.2829400| 0.6484187|
#> |E_004 | 7966.273| 2311.057| 0.0212316|  8.036047| 1039.098| 39.13434| 36.66475| 0.2045208| 0.5553003|
#> |E_005 | 7514.089| 2315.675| 0.0233063|  9.482051| 1091.607| 40.22041| 43.24368| 0.2010852| 0.6602820|
#> |E_006 | 7608.464| 1995.193| 0.0588128| 29.329605| 1096.871| 75.23614| 53.64705| 0.3534595| 0.1035969|