Check variable types

Check variable types in a data frame. Useful whenever the data must be filter prior some numerical analysis or procwessing

check_types(raw_data, excluded_columns = NULL, numeric = TRUE, quiet = TRUE)

Arguments

raw_data
data frame containing the raw data

excluded_columns
list of excluded columns (won't be check)

numeric
boolean to indicate the operation mode. If TRUE will look up for non-numeric variables and exclude them, otherwise it will filter numeric variables, including those with the following classes: - complex e.g. 1+i - integer e.g. 1L - numeric e.g. 1, 1.01

quiet
boolean to hide warning messages

Value

vector with column indices not meeting the checking mode, if excluded_columns != NULL, returns a vector with both originally excluded and newly found column indices

Examples

example_data <- data.frame(ID = c(1,2,3,4,5), 
                           P1 = c("one", "two", "three", "four", "five"), 
                           T1 = rnorm(5))
MetaPipe::check_types(example_data)

#> [1] 2

MetaPipe::check_types(example_data, numeric = FALSE)

#> [1] 1 3

MetaPipe::check_types(example_data, excluded_columns = 1, numeric = FALSE)

#> [1] 1 2

MetaPipe::check_types(example_data, quiet = FALSE)

#> Warning: 
#> The following variable will be excluded, as it is non-numeric: 
#>  - P1

#> [1] 2