Power transformation

Normalise data with a power transformation.

power_transformation(
  data,
  trait = "DATA",
  alpha = 0.05,
  transf_vals = c(2, exp(1), 3, 4, 5, 6, 7, 8, 9, 10),
  plots_prefix = "HIST",
  digits = 6,
  plot = TRUE,
  quiet = FALSE
)

Arguments

data
Original data.

trait
Trait name.

alpha
Significance level.

transf_vals
Transformation values.

plots_prefix
Prefix for plots with or without path.

digits
Significant digits to compare p-values of transformations.

plot
Boolean flag to indicate whether or not to plot a histogram comparing the original and transformed/normalised data.

quiet
Boolean flag to hide status messages.

Value

Data structure containing the normalised data, original data if no transformation was performed.

See also

Examples

set.seed(123)
data <- rnorm(100, 5)
MetaPipe::power_transformation(sqrt(data), "ROOT_2")

#> Warning: failed to load cairo DLL

#> 
#> The data was normalised by a power of 2

#>   [1] 4.439524 4.769823 6.558708 5.070508 5.129288 6.715065 5.460916 3.734939
#>   [9] 4.313147 4.554338 6.224082 5.359814 5.400771 5.110683 4.444159 6.786913
#>  [17] 5.497850 3.033383 5.701356 4.527209 3.932176 4.782025 3.973996 4.271109
#>  [25] 4.374961 3.313307 5.837787 5.153373 3.861863 6.253815 5.426464 4.704929
#>  [33] 5.895126 5.878133 5.821581 5.688640 5.553918 4.938088 4.694037 4.619529
#>  [41] 4.305293 4.792083 3.734604 7.168956 6.207962 3.876891 4.597115 4.533345
#>  [49] 5.779965 4.916631 5.253319 4.971453 4.957130 6.368602 4.774229 6.516471
#>  [57] 3.451247 5.584614 5.123854 5.215942 5.379639 4.497677 4.666793 3.981425
#>  [65] 3.928209 5.303529 5.448210 5.053004 5.922267 7.050085 4.508969 2.690831
#>  [73] 6.005739 4.290799 4.311991 6.025571 4.715227 3.779282 5.181303 4.861109
#>  [81] 5.005764 5.385280 4.629340 5.644377 4.779513 5.331782 6.096839 5.435181
#>  [89] 4.674068 6.148808 5.993504 5.548397 5.238732 4.372094 6.360652 4.399740
#>  [97] 7.187333 6.532611 4.764300 3.973579