Read QTL data

Read QTL data from two sources, one containing genotypes and another one phenotypes.

read.cross(geno, pheno, wdir = here::here(), quiet = TRUE, ...)

Arguments

geno
Data frame or string (filename) to file containing genotype data. For example a genetic map like father_riparia.

pheno
Data frame or string (filename) to file containing phenotype data. For example, output from assess_normality.

wdir
Working directory.

quiet
Boolean flag to hide status messages.

...
Arguments passed on to qtl::read.cross

file

The main input file for formats csv, csvr and mm.

genfile

File with genotype data (formats csvs, csvsr, karl, gary and mapqtl only).

mapfile

File with marker position information (all except the csv formats).

phefile

File with phenotype data (formats csvs, csvsr, karl, gary and mapqtl only).

chridfile

File with chromosome ID for each marker (gary format only).

mnamesfile

File with marker names (gary format only).

pnamesfile

File with phenotype names (gary format only).

na.strings

A vector of strings which are to be interpreted as missing values (csv and gary formats only). For the csv formats, these are interpreted globally for the entire file, so missing value codes in phenotypes must not be valid genotypes, and vice versa. For the gary format, these are used only for the phenotype data.

genotypes

A vector of character strings specifying the genotype codes (csv formats only). Generally this is a vector of length 5, with the elements corresponding to AA, AB, BB, not BB (i.e., AA or AB), and not AA (i.e., AB or BB). Note: Pay careful attention to the third and fourth of these; the order of these can be confusing. If you are trying to read 4-way cross data, your file must have genotypes coded as described below, and you need to set genotypes=NULL so that no re-coding gets done.

alleles

A vector of two one-letter character strings (or four, for the four-way cross), to be used as labels for the two alleles.

estimate.map

For all formats but qtlcart, mapqtl, and karl: if TRUE and marker positions are not included in the input files, the genetic map is estimated using the function est.map.

convertXdata

If TRUE, any X chromosome genotype data is converted to the internal standard, using columns sex and pgm in the phenotype data if they available or by inference if they are not. If FALSE, the X chromsome data is read as is.

error.prob

In the case that the marker map must be estimated: Assumed genotyping error rate used in the calculation of the penetrance Pr(observed genotype | true genotype).

map.function

In the case that the marker map must be estimated: Indicates whether to use the Haldane, Kosambi, Carter-Falconer, or Morgan map function when converting genetic distances into recombination fractions. (Ignored if m > 0.)

BC.gen

Used only for cross type "bcsft".

F.gen

Used only for cross type "bcsft".

crosstype

Optional character string to force a particular cross type.

Value

Object of cross class for QTL mapping.

See also

Examples

data(father_riparia)
data(ionomics)
ionomics_rev <- MetaPipe::replace_missing(ionomics, 
                                          excluded_columns = c(1, 2),
                                          replace_na =  TRUE)
ionomics_normalised <- 
  MetaPipe::assess_normality(ionomics_rev,
                             excluded_columns = c(1, 2),
                             out_prefix = "ionomics",
                             transf_vals = c(2, exp(1)),
                             show_stats = FALSE)
x_data <- MetaPipe::read.cross(father_riparia, 
                               ionomics_normalised$norm,
                               genotypes = c("nn", "np", "--"))

#>  --Read the following data:
#> 	 166  individuals
#> 	 1115  markers
#> 	 7  phenotypes

#> Warning: Some markers at the same position on chr 1,4,5,7,8,9,10,12,14,15,16,17; use jittermap().

#>  --Cross type: f2