Root transformation

Normalise data with a root transformation.

root_transformation(
  data,
  trait = "DATA",
  alpha = 0.05,
  transf_vals = c(2, exp(1), 3, 4, 5, 6, 7, 8, 9, 10),
  plots_prefix = "HIST",
  digits = 6,
  plot = TRUE,
  quiet = FALSE
)

Arguments

data
Original data.

trait
Trait name.

alpha
Significance level.

transf_vals
Transformation values.

plots_prefix
Prefix for plots with or without path.

digits
Significant digits to compare p-values of transformations.

plot
Boolean flag to indicate whether or not to plot a histogram comparing the original and transformed/normalised data.

quiet
Boolean flag to hide status messages.

Value

Data structure containing the normalised data, original data if no transformation was performed.

See also

Examples

set.seed(123)
data <- rnorm(100, 5)
MetaPipe::root_transformation(data ^ 2, "EXP_2")

#> Warning: failed to load cairo DLL

#> 
#> The data was normalised by a root of e

#>   [1] 2.994219 3.156560 3.990078 3.301778 3.329897 4.059846 3.486982 2.636712
#>   [9] 2.931268 3.051000 3.839264 3.439366 3.458684 3.321006 2.996518 4.091761
#>  [17] 3.504318 2.262467 3.599297 3.037618 2.738461 3.162499 2.759859 2.910220
#>  [25] 2.962119 2.414277 3.662469 3.341394 2.702346 3.852750 3.470782 3.124905
#>  [33] 3.688902 3.681076 3.654986 3.593389 3.530577 3.238114 3.119581 3.083072
#>  [41] 2.927340 3.167392 2.636538 4.259997 3.831945 2.710080 3.072059 3.040646
#>  [49] 3.635744 3.227755 3.388953 3.254197 3.247296 3.904654 3.158705 3.971155
#>  [57] 2.487830 3.544923 3.327301 3.371195 3.448722 3.023026 3.106249 2.763654
#>  [65] 2.736427 3.412755 3.481010 3.293388 3.701391 4.207911 3.028608 2.071537
#>  [73] 3.739703 2.920086 2.930690 3.748786 3.129936 2.659709 3.354709 3.200896
#>  [81] 3.270706 3.451382 3.087888 3.572795 3.161277 3.426122 3.781358 3.474884
#>  [89] 3.109812 3.805046 3.734097 3.527994 3.382027 2.960690 3.901068 2.974453
#>  [97] 4.268029 3.978390 3.153870 2.759646