Function to save a graphical object to disk in Portable Document Format (PDF) format.

Function to save a graphical object to disk in Portable Document Format (PDF) format.

save_plotPDF(myPlot, name, width = 6, height = 6)

Arguments

myPlot
graphical object

name
output name with or without path

width
width in inches

height
height in inches

See also

Examples

save_plotPDF(hist(rnorm(100), main = "Histogram of Normal Distribution"),
"hist")

#> $breaks
#>  [1] -2.0 -1.5 -1.0 -0.5  0.0  0.5  1.0  1.5  2.0  2.5
#> 
#> $counts
#> [1]  3  8 17 20 21 10 10  8  3
#> 
#> $density
#> [1] 0.06 0.16 0.34 0.40 0.42 0.20 0.20 0.16 0.06
#> 
#> $mids
#> [1] -1.75 -1.25 -0.75 -0.25  0.25  0.75  1.25  1.75  2.25
#> 
#> $xname
#> [1] "rnorm(100)"
#> 
#> $equidist
#> [1] TRUE
#> 
#> attr(,"class")
#> [1] "histogram"

#> pdf 
#>   2