Normalise data using different methods: log, power, and root, with a different number of parameters to find out which one transforms the data into a normal-ish set.

Normalise data using different methods: log, power, and root, with a different number of parameters to find out which one transforms the data into a normal-ish set.

transform_data(
  data,
  feature = "DATA",
  alpha = 0.05,
  index = NULL,
  transf_vals = c(2, exp(1), 3, 4, 5, 6, 7, 8, 9, 10),
  plots_prefix = "HIST",
  digits = 6
)

Arguments

data
original data

feature
feature name

alpha
significance level

index
index of the current feature

transf_vals
transformation values

plots_prefix
prefix for plots with or without path

digits
significant digits to compare p-values of transformations

Value

transformed record

Examples

set.seed(123)
data <- rnorm(100, 5)
transform_data(2 ^ data, "EXP_2")

#> Warning: failed to load cairo DLL

#>     index feature   values   flag transf transf_val
#> 1           EXP_2 4.439524 Normal    log          2
#> 2           EXP_2 4.769823 Normal    log          2
#> 3           EXP_2 6.558708 Normal    log          2
#> 4           EXP_2 5.070508 Normal    log          2
#> 5           EXP_2 5.129288 Normal    log          2
#> 6           EXP_2 6.715065 Normal    log          2
#> 7           EXP_2 5.460916 Normal    log          2
#> 8           EXP_2 3.734939 Normal    log          2
#> 9           EXP_2 4.313147 Normal    log          2
#> 10          EXP_2 4.554338 Normal    log          2
#> 11          EXP_2 6.224082 Normal    log          2
#> 12          EXP_2 5.359814 Normal    log          2
#> 13          EXP_2 5.400771 Normal    log          2
#> 14          EXP_2 5.110683 Normal    log          2
#> 15          EXP_2 4.444159 Normal    log          2
#> 16          EXP_2 6.786913 Normal    log          2
#> 17          EXP_2 5.497850 Normal    log          2
#> 18          EXP_2 3.033383 Normal    log          2
#> 19          EXP_2 5.701356 Normal    log          2
#> 20          EXP_2 4.527209 Normal    log          2
#> 21          EXP_2 3.932176 Normal    log          2
#> 22          EXP_2 4.782025 Normal    log          2
#> 23          EXP_2 3.973996 Normal    log          2
#> 24          EXP_2 4.271109 Normal    log          2
#> 25          EXP_2 4.374961 Normal    log          2
#> 26          EXP_2 3.313307 Normal    log          2
#> 27          EXP_2 5.837787 Normal    log          2
#> 28          EXP_2 5.153373 Normal    log          2
#> 29          EXP_2 3.861863 Normal    log          2
#> 30          EXP_2 6.253815 Normal    log          2
#> 31          EXP_2 5.426464 Normal    log          2
#> 32          EXP_2 4.704929 Normal    log          2
#> 33          EXP_2 5.895126 Normal    log          2
#> 34          EXP_2 5.878133 Normal    log          2
#> 35          EXP_2 5.821581 Normal    log          2
#> 36          EXP_2 5.688640 Normal    log          2
#> 37          EXP_2 5.553918 Normal    log          2
#> 38          EXP_2 4.938088 Normal    log          2
#> 39          EXP_2 4.694037 Normal    log          2
#> 40          EXP_2 4.619529 Normal    log          2
#> 41          EXP_2 4.305293 Normal    log          2
#> 42          EXP_2 4.792083 Normal    log          2
#> 43          EXP_2 3.734604 Normal    log          2
#> 44          EXP_2 7.168956 Normal    log          2
#> 45          EXP_2 6.207962 Normal    log          2
#> 46          EXP_2 3.876891 Normal    log          2
#> 47          EXP_2 4.597115 Normal    log          2
#> 48          EXP_2 4.533345 Normal    log          2
#> 49          EXP_2 5.779965 Normal    log          2
#> 50          EXP_2 4.916631 Normal    log          2
#> 51          EXP_2 5.253319 Normal    log          2
#> 52          EXP_2 4.971453 Normal    log          2
#> 53          EXP_2 4.957130 Normal    log          2
#> 54          EXP_2 6.368602 Normal    log          2
#> 55          EXP_2 4.774229 Normal    log          2
#> 56          EXP_2 6.516471 Normal    log          2
#> 57          EXP_2 3.451247 Normal    log          2
#> 58          EXP_2 5.584614 Normal    log          2
#> 59          EXP_2 5.123854 Normal    log          2
#> 60          EXP_2 5.215942 Normal    log          2
#> 61          EXP_2 5.379639 Normal    log          2
#> 62          EXP_2 4.497677 Normal    log          2
#> 63          EXP_2 4.666793 Normal    log          2
#> 64          EXP_2 3.981425 Normal    log          2
#> 65          EXP_2 3.928209 Normal    log          2
#> 66          EXP_2 5.303529 Normal    log          2
#> 67          EXP_2 5.448210 Normal    log          2
#> 68          EXP_2 5.053004 Normal    log          2
#> 69          EXP_2 5.922267 Normal    log          2
#> 70          EXP_2 7.050085 Normal    log          2
#> 71          EXP_2 4.508969 Normal    log          2
#> 72          EXP_2 2.690831 Normal    log          2
#> 73          EXP_2 6.005739 Normal    log          2
#> 74          EXP_2 4.290799 Normal    log          2
#> 75          EXP_2 4.311991 Normal    log          2
#> 76          EXP_2 6.025571 Normal    log          2
#> 77          EXP_2 4.715227 Normal    log          2
#> 78          EXP_2 3.779282 Normal    log          2
#> 79          EXP_2 5.181303 Normal    log          2
#> 80          EXP_2 4.861109 Normal    log          2
#> 81          EXP_2 5.005764 Normal    log          2
#> 82          EXP_2 5.385280 Normal    log          2
#> 83          EXP_2 4.629340 Normal    log          2
#> 84          EXP_2 5.644377 Normal    log          2
#> 85          EXP_2 4.779513 Normal    log          2
#> 86          EXP_2 5.331782 Normal    log          2
#> 87          EXP_2 6.096839 Normal    log          2
#> 88          EXP_2 5.435181 Normal    log          2
#> 89          EXP_2 4.674068 Normal    log          2
#> 90          EXP_2 6.148808 Normal    log          2
#> 91          EXP_2 5.993504 Normal    log          2
#> 92          EXP_2 5.548397 Normal    log          2
#> 93          EXP_2 5.238732 Normal    log          2
#> 94          EXP_2 4.372094 Normal    log          2
#> 95          EXP_2 6.360652 Normal    log          2
#> 96          EXP_2 4.399740 Normal    log          2
#> 97          EXP_2 7.187333 Normal    log          2
#> 98          EXP_2 6.532611 Normal    log          2
#> 99          EXP_2 4.764300 Normal    log          2
#> 100         EXP_2 3.973579 Normal    log          2

transform_data(sqrt(data), "ROOT_2")

#> Warning: failed to load cairo DLL

#>     index feature   values   flag transf transf_val
#> 1          ROOT_2 4.439524 Normal  power          2
#> 2          ROOT_2 4.769823 Normal  power          2
#> 3          ROOT_2 6.558708 Normal  power          2
#> 4          ROOT_2 5.070508 Normal  power          2
#> 5          ROOT_2 5.129288 Normal  power          2
#> 6          ROOT_2 6.715065 Normal  power          2
#> 7          ROOT_2 5.460916 Normal  power          2
#> 8          ROOT_2 3.734939 Normal  power          2
#> 9          ROOT_2 4.313147 Normal  power          2
#> 10         ROOT_2 4.554338 Normal  power          2
#> 11         ROOT_2 6.224082 Normal  power          2
#> 12         ROOT_2 5.359814 Normal  power          2
#> 13         ROOT_2 5.400771 Normal  power          2
#> 14         ROOT_2 5.110683 Normal  power          2
#> 15         ROOT_2 4.444159 Normal  power          2
#> 16         ROOT_2 6.786913 Normal  power          2
#> 17         ROOT_2 5.497850 Normal  power          2
#> 18         ROOT_2 3.033383 Normal  power          2
#> 19         ROOT_2 5.701356 Normal  power          2
#> 20         ROOT_2 4.527209 Normal  power          2
#> 21         ROOT_2 3.932176 Normal  power          2
#> 22         ROOT_2 4.782025 Normal  power          2
#> 23         ROOT_2 3.973996 Normal  power          2
#> 24         ROOT_2 4.271109 Normal  power          2
#> 25         ROOT_2 4.374961 Normal  power          2
#> 26         ROOT_2 3.313307 Normal  power          2
#> 27         ROOT_2 5.837787 Normal  power          2
#> 28         ROOT_2 5.153373 Normal  power          2
#> 29         ROOT_2 3.861863 Normal  power          2
#> 30         ROOT_2 6.253815 Normal  power          2
#> 31         ROOT_2 5.426464 Normal  power          2
#> 32         ROOT_2 4.704929 Normal  power          2
#> 33         ROOT_2 5.895126 Normal  power          2
#> 34         ROOT_2 5.878133 Normal  power          2
#> 35         ROOT_2 5.821581 Normal  power          2
#> 36         ROOT_2 5.688640 Normal  power          2
#> 37         ROOT_2 5.553918 Normal  power          2
#> 38         ROOT_2 4.938088 Normal  power          2
#> 39         ROOT_2 4.694037 Normal  power          2
#> 40         ROOT_2 4.619529 Normal  power          2
#> 41         ROOT_2 4.305293 Normal  power          2
#> 42         ROOT_2 4.792083 Normal  power          2
#> 43         ROOT_2 3.734604 Normal  power          2
#> 44         ROOT_2 7.168956 Normal  power          2
#> 45         ROOT_2 6.207962 Normal  power          2
#> 46         ROOT_2 3.876891 Normal  power          2
#> 47         ROOT_2 4.597115 Normal  power          2
#> 48         ROOT_2 4.533345 Normal  power          2
#> 49         ROOT_2 5.779965 Normal  power          2
#> 50         ROOT_2 4.916631 Normal  power          2
#> 51         ROOT_2 5.253319 Normal  power          2
#> 52         ROOT_2 4.971453 Normal  power          2
#> 53         ROOT_2 4.957130 Normal  power          2
#> 54         ROOT_2 6.368602 Normal  power          2
#> 55         ROOT_2 4.774229 Normal  power          2
#> 56         ROOT_2 6.516471 Normal  power          2
#> 57         ROOT_2 3.451247 Normal  power          2
#> 58         ROOT_2 5.584614 Normal  power          2
#> 59         ROOT_2 5.123854 Normal  power          2
#> 60         ROOT_2 5.215942 Normal  power          2
#> 61         ROOT_2 5.379639 Normal  power          2
#> 62         ROOT_2 4.497677 Normal  power          2
#> 63         ROOT_2 4.666793 Normal  power          2
#> 64         ROOT_2 3.981425 Normal  power          2
#> 65         ROOT_2 3.928209 Normal  power          2
#> 66         ROOT_2 5.303529 Normal  power          2
#> 67         ROOT_2 5.448210 Normal  power          2
#> 68         ROOT_2 5.053004 Normal  power          2
#> 69         ROOT_2 5.922267 Normal  power          2
#> 70         ROOT_2 7.050085 Normal  power          2
#> 71         ROOT_2 4.508969 Normal  power          2
#> 72         ROOT_2 2.690831 Normal  power          2
#> 73         ROOT_2 6.005739 Normal  power          2
#> 74         ROOT_2 4.290799 Normal  power          2
#> 75         ROOT_2 4.311991 Normal  power          2
#> 76         ROOT_2 6.025571 Normal  power          2
#> 77         ROOT_2 4.715227 Normal  power          2
#> 78         ROOT_2 3.779282 Normal  power          2
#> 79         ROOT_2 5.181303 Normal  power          2
#> 80         ROOT_2 4.861109 Normal  power          2
#> 81         ROOT_2 5.005764 Normal  power          2
#> 82         ROOT_2 5.385280 Normal  power          2
#> 83         ROOT_2 4.629340 Normal  power          2
#> 84         ROOT_2 5.644377 Normal  power          2
#> 85         ROOT_2 4.779513 Normal  power          2
#> 86         ROOT_2 5.331782 Normal  power          2
#> 87         ROOT_2 6.096839 Normal  power          2
#> 88         ROOT_2 5.435181 Normal  power          2
#> 89         ROOT_2 4.674068 Normal  power          2
#> 90         ROOT_2 6.148808 Normal  power          2
#> 91         ROOT_2 5.993504 Normal  power          2
#> 92         ROOT_2 5.548397 Normal  power          2
#> 93         ROOT_2 5.238732 Normal  power          2
#> 94         ROOT_2 4.372094 Normal  power          2
#> 95         ROOT_2 6.360652 Normal  power          2
#> 96         ROOT_2 4.399740 Normal  power          2
#> 97         ROOT_2 7.187333 Normal  power          2
#> 98         ROOT_2 6.532611 Normal  power          2
#> 99         ROOT_2 4.764300 Normal  power          2
#> 100        ROOT_2 3.973579 Normal  power          2

transform_data(data ^ 2, "POW_2")

#> Warning: failed to load cairo DLL

#>     index feature   values   flag transf transf_val
#> 1           POW_2 2.994219 Normal   root          e
#> 2           POW_2 3.156560 Normal   root          e
#> 3           POW_2 3.990078 Normal   root          e
#> 4           POW_2 3.301778 Normal   root          e
#> 5           POW_2 3.329897 Normal   root          e
#> 6           POW_2 4.059846 Normal   root          e
#> 7           POW_2 3.486982 Normal   root          e
#> 8           POW_2 2.636712 Normal   root          e
#> 9           POW_2 2.931268 Normal   root          e
#> 10          POW_2 3.051000 Normal   root          e
#> 11          POW_2 3.839264 Normal   root          e
#> 12          POW_2 3.439366 Normal   root          e
#> 13          POW_2 3.458684 Normal   root          e
#> 14          POW_2 3.321006 Normal   root          e
#> 15          POW_2 2.996518 Normal   root          e
#> 16          POW_2 4.091761 Normal   root          e
#> 17          POW_2 3.504318 Normal   root          e
#> 18          POW_2 2.262467 Normal   root          e
#> 19          POW_2 3.599297 Normal   root          e
#> 20          POW_2 3.037618 Normal   root          e
#> 21          POW_2 2.738461 Normal   root          e
#> 22          POW_2 3.162499 Normal   root          e
#> 23          POW_2 2.759859 Normal   root          e
#> 24          POW_2 2.910220 Normal   root          e
#> 25          POW_2 2.962119 Normal   root          e
#> 26          POW_2 2.414277 Normal   root          e
#> 27          POW_2 3.662469 Normal   root          e
#> 28          POW_2 3.341394 Normal   root          e
#> 29          POW_2 2.702346 Normal   root          e
#> 30          POW_2 3.852750 Normal   root          e
#> 31          POW_2 3.470782 Normal   root          e
#> 32          POW_2 3.124905 Normal   root          e
#> 33          POW_2 3.688902 Normal   root          e
#> 34          POW_2 3.681076 Normal   root          e
#> 35          POW_2 3.654986 Normal   root          e
#> 36          POW_2 3.593389 Normal   root          e
#> 37          POW_2 3.530577 Normal   root          e
#> 38          POW_2 3.238114 Normal   root          e
#> 39          POW_2 3.119581 Normal   root          e
#> 40          POW_2 3.083072 Normal   root          e
#> 41          POW_2 2.927340 Normal   root          e
#> 42          POW_2 3.167392 Normal   root          e
#> 43          POW_2 2.636538 Normal   root          e
#> 44          POW_2 4.259997 Normal   root          e
#> 45          POW_2 3.831945 Normal   root          e
#> 46          POW_2 2.710080 Normal   root          e
#> 47          POW_2 3.072059 Normal   root          e
#> 48          POW_2 3.040646 Normal   root          e
#> 49          POW_2 3.635744 Normal   root          e
#> 50          POW_2 3.227755 Normal   root          e
#> 51          POW_2 3.388953 Normal   root          e
#> 52          POW_2 3.254197 Normal   root          e
#> 53          POW_2 3.247296 Normal   root          e
#> 54          POW_2 3.904654 Normal   root          e
#> 55          POW_2 3.158705 Normal   root          e
#> 56          POW_2 3.971155 Normal   root          e
#> 57          POW_2 2.487830 Normal   root          e
#> 58          POW_2 3.544923 Normal   root          e
#> 59          POW_2 3.327301 Normal   root          e
#> 60          POW_2 3.371195 Normal   root          e
#> 61          POW_2 3.448722 Normal   root          e
#> 62          POW_2 3.023026 Normal   root          e
#> 63          POW_2 3.106249 Normal   root          e
#> 64          POW_2 2.763654 Normal   root          e
#> 65          POW_2 2.736427 Normal   root          e
#> 66          POW_2 3.412755 Normal   root          e
#> 67          POW_2 3.481010 Normal   root          e
#> 68          POW_2 3.293388 Normal   root          e
#> 69          POW_2 3.701391 Normal   root          e
#> 70          POW_2 4.207911 Normal   root          e
#> 71          POW_2 3.028608 Normal   root          e
#> 72          POW_2 2.071537 Normal   root          e
#> 73          POW_2 3.739703 Normal   root          e
#> 74          POW_2 2.920086 Normal   root          e
#> 75          POW_2 2.930690 Normal   root          e
#> 76          POW_2 3.748786 Normal   root          e
#> 77          POW_2 3.129936 Normal   root          e
#> 78          POW_2 2.659709 Normal   root          e
#> 79          POW_2 3.354709 Normal   root          e
#> 80          POW_2 3.200896 Normal   root          e
#> 81          POW_2 3.270706 Normal   root          e
#> 82          POW_2 3.451382 Normal   root          e
#> 83          POW_2 3.087888 Normal   root          e
#> 84          POW_2 3.572795 Normal   root          e
#> 85          POW_2 3.161277 Normal   root          e
#> 86          POW_2 3.426122 Normal   root          e
#> 87          POW_2 3.781358 Normal   root          e
#> 88          POW_2 3.474884 Normal   root          e
#> 89          POW_2 3.109812 Normal   root          e
#> 90          POW_2 3.805046 Normal   root          e
#> 91          POW_2 3.734097 Normal   root          e
#> 92          POW_2 3.527994 Normal   root          e
#> 93          POW_2 3.382027 Normal   root          e
#> 94          POW_2 2.960690 Normal   root          e
#> 95          POW_2 3.901068 Normal   root          e
#> 96          POW_2 2.974453 Normal   root          e
#> 97          POW_2 4.268029 Normal   root          e
#> 98          POW_2 3.978390 Normal   root          e
#> 99          POW_2 3.153870 Normal   root          e
#> 100         POW_2 2.759646 Normal   root          e