Read QTL data

Read QTL data from two sources, one containing genotypes and another one phenotypes.

Usage,

read.cross(geno, pheno, wdir = tempdir(), quiet = TRUE, ...)

Arguments

geno

Data frame or string (filename) to file containing genotype data. For example a genetic map like father_riparia.

pheno

Data frame or string (filename) to file containing phenotype data. For example, output from assess_normality.

wdir

Working directory.

quiet

Boolean flag to hide status messages.

...

Arguments passed on to qtl::read.cross.

Value

Object of cross class for QTL mapping.

See also

Examples

data(father_riparia)
data(ionomics)
ionomics_rev <- MetaPipe::replace_missing(ionomics, 
                                          excluded_columns = c(1, 2),
                                          replace_na =  TRUE)
# \donttest{
ionomics_normalised <- 
  MetaPipe::assess_normality(ionomics_rev,
                             excluded_columns = c(1, 2),
                             out_prefix = "ionomics",
                             transf_vals = c(2, exp(1)),
                             show_stats = FALSE)
x_data <- MetaPipe::read.cross(father_riparia, 
                               ionomics_normalised$norm,
                               genotypes = c("nn", "np", "--"))
#>  --Read the following data:
#> 	 166  individuals
#> 	 1115  markers
#> 	 7  phenotypes
#> Warning: Some markers at the same position on chr 1,4,5,7,8,9,10,12,14,15,16,17; use jittermap().
#>  --Cross type: f2 
# }