Replace missing values (NAs)

Replace missing values (NAs)

replace_missing(
  raw_data,
  excluded_columns,
  out_prefix = "metapipe",
  prop_na = 0.5,
  replace_na = FALSE
)

Arguments

raw_data
data frame containing the raw data

excluded_columns
vector containing the indices of the data set properties, excluded columns

out_prefix
prefix for output files, plots, etc.

prop_na
proportion of missing values, if a feature exceeds this threshold, then it is dropped out

replace_na
boolean flag to indicate whether or not missing values should be replaced by half of the minimum value

Value

data frame containing the raw data without missing values

Examples

example_data <- data.frame(ID = c(1,2,3,4,5),
                           P1 = c("one", "two", "three", "four", "five"),
                           T1 = rnorm(5),
                           T2 = rnorm(5),
                           T3 = c(NA, rnorm(4)),                  #  20 % NAs
                           T4 = c(NA, 1.2, -0.5, NA, 0.87),       #  40 % NAs
                           T5 = NA)                               # 100 % NAs
replace_missing(example_data, c(1, 2))

#> The following traits were dropped because they have 50% or more missing values: 
#> T5

#>   ID    P1          T1         T2         T3    T4
#> 1  1   one  0.20609871 -0.7505732         NA    NA
#> 2  2   two -0.09160289 -0.2871355  1.0464316  1.20
#> 3  3 three -0.46000136  0.8316341 -0.5196884 -0.50
#> 4  4  four  0.56749428 -0.5269867  1.0535059    NA
#> 5  5  five -1.46018697 -1.8326439  1.3615273  0.87

replace_missing(example_data, c(1, 2), prop_na =  0.25)

#> The following traits were dropped because they have 25% or more missing values: 
#> T4, T5

#>   ID    P1          T1         T2         T3
#> 1  1   one  0.20609871 -0.7505732         NA
#> 2  2   two -0.09160289 -0.2871355  1.0464316
#> 3  3 three -0.46000136  0.8316341 -0.5196884
#> 4  4  four  0.56749428 -0.5269867  1.0535059
#> 5  5  five -1.46018697 -1.8326439  1.3615273

replace_missing(example_data, c(1, 2), replace_na =  TRUE)

#> The following traits were dropped because they have 100% missing values: 
#> T5

#>   ID    P1          T1         T2         T3    T4
#> 1  1   one  0.20609871 -0.7505732 -0.2598442 -0.25
#> 2  2   two -0.09160289 -0.2871355  1.0464316  1.20
#> 3  3 three -0.46000136  0.8316341 -0.5196884 -0.50
#> 4  4  four  0.56749428 -0.5269867  1.0535059 -0.25
#> 5  5  five -1.46018697 -1.8326439  1.3615273  0.87