Replace missing values (NAs)

Replace missing values (NAs)

replace_missing(
  raw_data,
  excluded_columns,
  out_prefix = "metapipe",
  prop_na = 0.5,
  replace_na = FALSE
)

Arguments

raw_data
data frame containing the raw data

excluded_columns
vector containing the indices of the data set properties, excluded columns

out_prefix
prefix for output files, plots, etc.

prop_na
proportion of missing values, if a feature exceeds this threshold, then it is dropped out

replace_na
boolean flag to indicate whether or not missing values should be replaced by half of the minimum value

Value

data frame containing the raw data without missing values

Examples

example_data <- data.frame(ID = c(1,2,3,4,5),
                           P1 = c("one", "two", "three", "four", "five"),
                           F1 = rnorm(5),
                           F2 = rnorm(5))
example_data$F1[2:3] <- NA
example_data$F2[4] <- NA
replace_missing(example_data, c(1, 2))

#>   ID    P1         F1         F2
#> 1  1   one  0.2060987 -0.7505732
#> 2  2   two         NA -0.2871355
#> 3  3 three         NA  0.8316341
#> 4  4  four  0.5674943         NA
#> 5  5  five -1.4601870 -1.8326439

replace_missing(example_data, c(1, 2), prop_na =  0.25)

#> The following features were dropped because they have 25% or more missing values: F1

#>   ID    P1         F2
#> 1  1   one -0.7505732
#> 2  2   two -0.2871355
#> 3  3 three  0.8316341
#> 4  4  four         NA
#> 5  5  five -1.8326439

replace_missing(example_data, c(1, 2), replace_na =  TRUE)

#>   ID    P1         F1         F2
#> 1  1   one  0.2060987 -0.7505732
#> 2  2   two -0.7300935 -0.2871355
#> 3  3 three -0.7300935  0.8316341
#> 4  4  four  0.5674943 -0.9163220
#> 5  5  five -1.4601870 -1.8326439